+++ asn1c/tests/check-30.-fwide-types.c
@@ -142,7 +142,7 @@ uint8_t buf5[] = {
 };
 
 static void
-check(int is_ok, uint8_t *buf, int size, size_t consumed) {
+check(int is_ok, uint8_t *buf, size_t size, size_t consumed) {
 	T_t t, *tp;
 	asn_dec_rval_t rval;
 
@@ -152,8 +152,8 @@ check(int is_ok, uint8_t *buf, int size, size_t consumed) {
 
 	fprintf(stderr, "Buf %p\n", buf);
 	rval = ber_decode(0, &asn_DEF_T, (void **)&tp, buf, size);
-	fprintf(stderr, "Returned code %d, consumed %d (out of %d)\n",
-		(int)rval.code, (int)rval.consumed, size);
+	fprintf(stderr, "Returned code %d, consumed %zd (out of %zd)\n",
+		(int)rval.code, rval.consumed, size);
 
 	if(is_ok) {
 		assert(rval.code == RC_OK);
@@ -215,7 +215,7 @@ check_xer(uint8_t *buf, uint8_t size, char *xer_sample) {
 }
 
 static void
-try_corrupt(uint8_t *buf, int size) {
+try_corrupt(uint8_t *buf, size_t size) {
 	uint8_t *tmp;
 	int i;
