+++ asn1c/tests/check-35.c
@@ -123,15 +123,15 @@ uint8_t buf2_reconstr[] = {
 };
 
 static void
-check(T_t *tp, uint8_t *buf, int size, size_t consumed) {
+check(T_t *tp, uint8_t *buf, size_t size, size_t consumed) {
 	asn_dec_rval_t rval;
 
 	tp = memset(tp, 0, sizeof(*tp));
 
-	fprintf(stderr, "Buf %p (%d)\n", buf, (int)size);
+	fprintf(stderr, "Buf %p (%zd)\n", buf, size);
 	rval = ber_decode(0, &asn_DEF_T, (void **)&tp, buf, size);
-	fprintf(stderr, "Returned code %d, consumed %d\n",
-		(int)rval.code, (int)rval.consumed);
+	fprintf(stderr, "Returned code %d, consumed %zd\n",
+		(int)rval.code, rval.consumed);
 
 	assert(rval.code == RC_OK);
 	assert(rval.consumed == consumed);
@@ -154,20 +154,20 @@ buf_fill(const void *buffer, size_t size, void *app_key) {
 	(void)app_key;
 
 	if(buf_pos + size > buf_size) {
-		fprintf(stderr, "%d + %d > %d\n",
-			(int)buf_pos, (int)size, (int)buf_size);
+		fprintf(stderr, "%zd + %zd > %zd\n",
+			buf_pos, size, buf_size);
 		return -1;
 	}
 
 	memcpy(buf + buf_pos, buffer, size);
 	buf_pos += size;
-	fprintf(stderr, "   written %d (%d)\n", (int)size, (int)buf_pos);
+	fprintf(stderr, "   written %zd (%zd)\n", size, buf_pos);
 
 	return 0;
 }
 
 static void
-compare(T_t *tp, uint8_t *cmp_buf, int cmp_buf_size) {
+compare(T_t *tp, uint8_t *cmp_buf, size_t cmp_buf_size) {
 	asn_enc_rval_t erval;
 	int i;
 
@@ -181,7 +181,7 @@ compare(T_t *tp, uint8_t *cmp_buf, int cmp_buf_size) {
 	erval = der_encode(&asn_DEF_T, tp, buf_fill, 0);
 	assert(erval.encoded != -1);
 	if(erval.encoded != cmp_buf_size) {
-		printf("%d != %d\n", erval.encoded, cmp_buf_size);
+		printf("%zd != %zd\n", erval.encoded, cmp_buf_size);
 	}
 	assert(erval.encoded == cmp_buf_size);
 	for(i = 0; i < cmp_buf_size; i++) {
@@ -223,8 +223,8 @@ partial_read(uint8_t *buf, size_t size) {
 			uint8_t *chunk3 = buf + size1 + size2;
 			size_t size3 = size - size1 - size2;
 
-			fprintf(stderr, "\n%d:{%d, %d, %d}...\n",
-				(int)size, (int)size1, (int)size2, (int)size3);
+			fprintf(stderr, "\n%zd:{%zd, %zd, %zd}...\n",
+				size, size1, size2, size3);
 
 			memset(buf1, 0, size);
 			memset(buf2, 0, size);
@@ -235,7 +235,7 @@ partial_read(uint8_t *buf, size_t size) {
 
 			tp = memset(&t, 0, sizeof(t));
 
-			fprintf(stderr, "=> Chunk 1 (%d):\n", (int)size1);
+			fprintf(stderr, "=> Chunk 1 (%zd):\n", size1);
 			rval = ber_decode(0, &asn_DEF_T, (void **)&tp,
 				buf1, size1);
 			assert(rval.code == RC_WMORE);
@@ -247,7 +247,7 @@ partial_read(uint8_t *buf, size_t size) {
 				size2 += leftover;
 			}
 
-			fprintf(stderr, "=> Chunk 2 (%d):\n", (int)size2);
+			fprintf(stderr, "=> Chunk 2 (%zd):\n", size2);
 			rval = ber_decode(0, &asn_DEF_T, (void **)&tp,
 				buf2, size2);
 			assert(rval.code == RC_WMORE);
@@ -259,7 +259,7 @@ partial_read(uint8_t *buf, size_t size) {
 				size3 += leftover;
 			}
 
-			fprintf(stderr, "=> Chunk 3 (%d):\n", (int)size3);
+			fprintf(stderr, "=> Chunk 3 (%zd):\n", size3);
 			rval = ber_decode(0, &asn_DEF_T, (void **)&tp,
 				buf3, size3);
 			assert(rval.code == RC_OK);
@@ -271,7 +271,7 @@ partial_read(uint8_t *buf, size_t size) {
 }
 
 static char xer_buf[128];
-static int xer_off;
+static size_t xer_off;
 
 static int
 xer_cb(const void *buffer, size_t size, void *key) {
@@ -287,7 +287,7 @@ check_xer(uint8_t *buf, uint8_t size, char *xer_sample) {
 	T_t *tp = 0;
 	asn_dec_rval_t rval;
 	asn_enc_rval_t er;
-	int xer_sample_len = strlen(xer_sample);
+	size_t xer_sample_len = strlen(xer_sample);
 
 	rval = ber_decode(0, &asn_DEF_T, (void **)&tp, buf, size);
 	assert(rval.code == RC_OK);
@@ -298,7 +298,7 @@ check_xer(uint8_t *buf, uint8_t size, char *xer_sample) {
 	er = xer_encode(&asn_DEF_T, tp, XER_F_CANONICAL, xer_cb, 0);
 	assert(xer_off);
 	xer_buf[xer_off] = 0;
-	printf("[%s] (%d/%d) vs [%s] (%d)\n",
+	printf("[%s] (%zd/%zd) vs [%s] (%zd)\n",
 		xer_buf, er.encoded, xer_off, xer_sample, xer_sample_len);
 	assert(er.encoded == xer_off);
 	assert(xer_off == xer_sample_len);
